requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pyyaml>=6.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# LangChain + LangGraph (agent orchestration)
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-google-vertexai>=2.0.0

# Google Cloud (Vertex AI + Storage + Secret Manager)
google-cloud-aiplatform>=1.65.0
google-cloud-storage>=2.10.0
google-cloud-secret-manager>=2.16.0

# Async support
aiofiles>=24.0.0
httpx>=0.24.0

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Development dependencies
black>=23.0.0
ruff>=0.1.0
mypy>=1.7.0
